hamburger menu functionality complete.
