+++ client/src/Sidebar.jsx
@@ -1,16 +1,12 @@
-import React, {useState} from 'react';
+import React from 'react';
 import { slide as Menu } from 'react-burger-menu';
 import "./Sidebar.css";
-import { Link } from 'react-router-dom';
-import { useNavigate } from 'react-router-dom';
-import Cookies from 'universal-cookie';
+
 // import LoggedInUser from '../context/AuthContext';
-import axios from 'axios';
 
 export default function Sidebar(props)  {
 
-  const cookies = new Cookies();
-  let user = document.cookie;
+  
 
 
   return (
@@ -18,9 +14,13 @@ export default function Sidebar(props)  {
       <a className="menu-item" href="/">
         Home
       </a>
-      <a  className="menu-item" href='/signin'>
+      {!props.user &&<a className="menu-item" href='/signin'>
         Log In
-      </a>
+      </a>}
+      {props.user &&<a className="menu-item" href='/' onClick={props.handleLogout}>
+        Log Out
+      </a>}
+      
       <a className="menu-item" href="/register">
         Register
       </a>
