+++ client/src/index.css
@@ -10,7 +10,7 @@
 
 body {
   margin: 0 auto;
-  max-width: 960px;
+  /* max-width: 960px; */
   font-family: "Roboto";
 }
